+++ src/main/kotlin/sndl/parnas/storage/impl/SSM.kt
@@ -1,37 +1,42 @@
 package sndl.parnas.storage.impl
 
-import com.amazonaws.auth.AWSCredentialsProviderChain
-import com.amazonaws.auth.EC2ContainerCredentialsProviderWrapper
-import com.amazonaws.auth.EnvironmentVariableCredentialsProvider
-import com.amazonaws.auth.profile.ProfileCredentialsProvider
-import com.amazonaws.regions.DefaultAwsRegionProviderChain
-import com.amazonaws.services.simplesystemsmanagement.AWSSimpleSystemsManagement
-import com.amazonaws.services.simplesystemsmanagement.AWSSimpleSystemsManagementClientBuilder
-import com.amazonaws.services.simplesystemsmanagement.model.*
 import sndl.parnas.storage.Storage
 import sndl.parnas.storage.ConfigOption
 import sndl.parnas.utils.*
+import software.amazon.awssdk.auth.credentials.AwsCredentialsProviderChain
+import software.amazon.awssdk.auth.credentials.EnvironmentVariableCredentialsProvider
+import software.amazon.awssdk.auth.credentials.InstanceProfileCredentialsProvider
+import software.amazon.awssdk.auth.credentials.ProfileCredentialsProvider
+import software.amazon.awssdk.regions.Region
+import software.amazon.awssdk.regions.providers.DefaultAwsRegionProviderChain
+import software.amazon.awssdk.services.ssm.SsmClient
+import software.amazon.awssdk.services.ssm.model.DeleteParameterRequest
+import software.amazon.awssdk.services.ssm.model.GetParameterRequest
+import software.amazon.awssdk.services.ssm.model.GetParametersByPathRequest
+import software.amazon.awssdk.services.ssm.model.ParameterNotFoundException
+import software.amazon.awssdk.services.ssm.model.PutParameterRequest
 import java.lang.reflect.InvocationHandler
 import java.lang.reflect.InvocationTargetException
 import java.lang.reflect.Method
 import java.lang.reflect.Proxy
 import kotlin.system.measureTimeMillis
 
-class SSM(name: String, ssmClient: AWSSimpleSystemsManagement,
+class SSM(name: String, ssmClient: SsmClient,
           var prefix: String, private val kmsKeyId: String?, private val separatorToReplace: String? = null) : Storage(name) {
 
     constructor(name: String, region: String?, profileName: String?,
                 prefix: String, keyId: String?, separatorToReplace: String? = null) :
             this(
                     name = name,
-                    ssmClient = AWSSimpleSystemsManagementClientBuilder.standard()
-                            .withRegion(region ?: DefaultAwsRegionProviderChain().region)
-                            .withCredentials(AWSCredentialsProviderChain(
-                                    EnvironmentVariableCredentialsProvider(),
-                                    profileName?.let { ProfileCredentialsProvider(it) },
-                                    EC2ContainerCredentialsProviderWrapper()
-                            ))
-                            .build(),
+                    ssmClient = SsmClient.builder()
+                            .region(region?.let { Region.of(it)} ?: DefaultAwsRegionProviderChain().region)
+                            .credentialsProvider(
+                                AwsCredentialsProviderChain.of(
+                                    EnvironmentVariableCredentialsProvider.create(),
+                                    profileName?.let { ProfileCredentialsProvider.create(it) },
+                                    InstanceProfileCredentialsProvider.create()
+                                )
+                            ).build(),
                     prefix = prefix,
                     kmsKeyId = keyId,
                     separatorToReplace = separatorToReplace
@@ -83,10 +88,10 @@ class SSM(name: String, ssmClient: AWSSimpleSystemsManagement,
     }
 
     private val ssmClientProxy = Proxy.newProxyInstance(
-            AWSSimpleSystemsManagement::class.java.classLoader,
-            arrayOf(AWSSimpleSystemsManagement::class.java),
-            SSMHandler(ssmClient)
-    ) as AWSSimpleSystemsManagement
+        SsmClient::class.java.classLoader,
+        arrayOf(SsmClient::class.java),
+        SSMHandler(ssmClient)
+    ) as SsmClient
 
     override val isInitialized: Boolean = true
 
@@ -102,31 +107,33 @@ class SSM(name: String, ssmClient: AWSSimpleSystemsManagement,
     }
 
     override fun list() = buildSet<ConfigOption> {
-        val request = GetParametersByPathRequest()
-                .withPath(prefix)
-                .withRecursive(true)
-                .withWithDecryption(true)
+        var request = GetParametersByPathRequest.builder()
+                .path(prefix)
+                .recursive(true)
+                .withDecryption(true)
+                .build()
 
         do {
             val result = ssmClientProxy.getParametersByPath(request)
-            val nextToken = result.nextToken
+            val nextToken = result.nextToken()
 
-            result.parameters.forEach {
-                add(ConfigOption(it.name.removePrefix(prefix).convertFromSSMFormat(), it.value))
+            result.parameters().forEach {
+                add(ConfigOption(it.name().removePrefix(prefix).convertFromSSMFormat(), it.value()))
             }
 
-            request.nextToken = nextToken
+            request = request.copy { it.nextToken(nextToken) }
         } while (nextToken != null)
     }
 
     override fun get(key: String): ConfigOption? {
         val fullKey = prefix + key.convertToSSMFormat()
-        val request = GetParameterRequest()
-                .withName(fullKey)
-                .withWithDecryption(true)
+        val request = GetParameterRequest.builder()
+                .name(fullKey)
+                .withDecryption(true)
+                .build()
 
         return try {
-            ConfigOption(key, ssmClientProxy.getParameter(request).parameter.value)
+            ConfigOption(key, ssmClientProxy.getParameter(request).parameter().value())
         } catch (e: ParameterNotFoundException) {
             null
         }
@@ -134,17 +141,17 @@ class SSM(name: String, ssmClient: AWSSimpleSystemsManagement,
 
     override fun set(key: String, value: String): ConfigOption {
         val fullKey = prefix + key.convertToSSMFormat()
-        val request = PutParameterRequest()
-                .withName(fullKey)
-                .withValue(value)
-                .withOverwrite(true)
+        val request = PutParameterRequest.builder()
+                .name(fullKey)
+                .value(value)
+                .overwrite(true)
                 .apply {
                     // TODO@sndl: support StringList type
                     kmsKeyId?.let {
-                        withType("SecureString")
-                        withKeyId(it)
-                    } ?: withType("String")
-                }
+                        type("SecureString")
+                        keyId(it)
+                    } ?: type("String")
+                }.build()
 
         ssmClientProxy.putParameter(request)
 
@@ -153,8 +160,9 @@ class SSM(name: String, ssmClient: AWSSimpleSystemsManagement,
 
     override fun delete(key: String) {
         val fullKey = prefix + key.convertToSSMFormat()
-        val request = DeleteParameterRequest()
-                .withName(fullKey)
+        val request = DeleteParameterRequest.builder()
+                .name(fullKey)
+                .build()
 
         try {
             ssmClientProxy.deleteParameter(request)