+++ src/test/kotlin/sndl/parnas/TestContainersFactory.kt
@@ -1,19 +1,16 @@
 package sndl.parnas
 
 import org.testcontainers.containers.GenericContainer
-import org.testcontainers.containers.wait.strategy.HttpWaitStrategy
+import org.testcontainers.containers.localstack.LocalStackContainer
+import org.testcontainers.utility.DockerImageName
 
 object TestContainersFactory {
     // See: https://github.com/testcontainers/testcontainers-java/issues/318
     class KGenericContainer(imageName: String) : GenericContainer<KGenericContainer>(imageName)
 
-    private const val localstackVersion = "0.12.1"
+    private const val LOCALSTACK_IMAGE = "localstack/localstack:3.0"
 
-    fun getLocalstack(services: String = "ssm"): KGenericContainer = KGenericContainer("localstack/localstack:$localstackVersion")
-            .withExposedPorts(4566)
-            .waitingFor(HttpWaitStrategy()
-                    .forPath("/health")
-                    .forStatusCode(200)
-                    .forResponsePredicate { it != "{}" })
-            .withEnv("SERVICES", services).also { it.start() }
+    fun getLocalstack(): LocalStackContainer = LocalStackContainer(DockerImageName.parse(LOCALSTACK_IMAGE))
+        .withServices(LocalStackContainer.Service.SSM)
+        .also { it.start() }
 }