+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.7
+* Dependency updates: Gradle 8.11.1, Kotlin 2.0.21, testcontainers with localstack 1.20.3/3.0.0
+* Migration from an outdated openkeepass library to KeePassJava
+* Migration to AWS SDK v2
+
 ## 0.2.5, 0.2.6
 * Minor version update for some libraries